Force server unit tests to run serially.
In case anyone gets fancy and tries to run the tests in parallel,
we want to make sure no server-side TestCases are run concurrently.
In general, parallel ctest running will be encouraged in girder
because we will have plenty of other tests that won't conflict,
e.g. front-end tests or style tests.