Added logging and topological sorting of plugins at load time
We added the girder.logger utility with split error and info logs
that automatically rotate. Also, we gave plugins the ability to
create a plugin.json file in their top level directory, which
will contain all of their declaration info such as name, desc,
and a list of other plugins they depend on.