Enables "one-shot" file upload in POST /file.
Introduces an optional parameter "contents" to the POST /file rest call, which,
if specified, must be smaller than a single chunk (implied to be 64MB elsewhere
in the codebase), and whose size must exactly match the required "size"
parameter.

If this is specified, the API makes the appropriate call to readChunk() with the
contents, as though the user had followed up the POST /file call with a single
POST /file/chunk call.