Changed the default api endpoint security to require admin access.  Each handler should be decorated with @admin, @user, or @anonymous to override the default (the default is the same as specifying @admin).
I have attempted to pick the right value for all of the endpoints, but it would be handy to have someone else review my choices.  I have removed some of the admin and user checks that would now be duplications of what the decorators do.  Some of the tests have changed, since the error messages are not necessarily the same.

I modified the existing loadmodel plugin to better presrve the function name using the functools.wraps function.  If we wanted to ensure that decorated functions and methods looked even more like their original functions, I'd recommend pulling in the decorator module with pip; this isn't required, but it makes some debugging of decorated functions and methods easier.