I had failed to page multipart uploads in the cancelUpload function.
While testing, I noticed that when the server_upload and server_file tests ran concurrently, the server_upload test would sometimes fail.  I'm not sure why this is, as it would fail in the S3 asseststore part of the test, but putting a resource lock so that these tests can run concurrently solves the problem as far as I can tell.