Sped up hierarchy widget paging on large sets.
I noticed when working with a large collection (>100000 items), that the hierarchy widget was very slow to load a page of data.  Internally, mongo was using the folderId index, but not the name index.  By adding an index that uses both of these fields, mongo is more sensible and pages the data quickly.