Changed deletes and downloads for checked resources.
This addresses bug #451 and adds feature #458.  Specifically, when deleting or downloading large numbers of items, the process would fail because the query string for the GET and DELETE methods would exceed allowed sizes.  The downloads have been switched to using a form submit process.  The deletes use a body on the rest request.  Since bodies on deletes are only partially allowed (they aren't supported via phantom and jasmine, as far as I can tell), I also added the ability to send a POST request with an X-HTTP-Method-Override header (or its variants) that allows using a POST and redirecting to a DELETE.  The testing was updated to reflect this.