Editing multiple files in an item didn't work.
On items with multiple files, the edit dialog was saved.  This means that when you edited the first file, all was well.  If you edited a second file without reloading the page, the dialog would should the first file again.

Mime-type is not a required field when creating a file.  However, if you had not specified it, you could not edit the file.