Added a default conversion for serializing JSON.
Also fixed an issue outputting a zip file with a large number of files.  Specifically, if the file count was greater than 65535 or the compressed size was larger than 2^32-1, the existing code would fail.

Add tests to check downloading metadata with a datetime and invloke some of the large zip file code.