Allow public access routes to be plain functions.
We had assumed all routes were methods of a subclass of Resource, but our documentation showed a plain function.  This now allows routes with public access to be plain functions and updates the docs to explain that other routes needs to be methods.  Also, fixed a few warnings in building the docs.