Refs #581. Invert child view registration behavior
Instead of making the parent view register the child views it
creates, we instead have child views attach themselves to the
parent when they are constructed. This behavior is modeled after
Qt, and allows us to issue console warnings when developers
forget to pass the parent view, so potential memory leaks will
be easy to spot.