Quota and assetstore policies for users and collections.
Added support for preferred and fallback assetstores on users and collections.

Added user interface for quota and assetstore policies for users and collections.  This includes pie charts for capacity.  Non admins can check their own user and collection capacities.

Implemented quotas for total file size for users and collections.

Added a GirderException so that errors that aren't strictly RestExceptions, are failures, but aren't catastrophic, can be handled distinctly.  For instance, if the assetstore for a particular user is unavailable, there is not necessarily anything wrong with the system (just with the assetstore).  These have an optional 'identifier' that can get passed through to a rest consumer or other error handler.  This obviates the need to parse the error text to determine the type of error.

Updated documentation for new events.

Parallelized web_client tests.  Most web_client tests can now be run in parallel.  Those involving monog replicasets cannot.  Those that upload via the phantomjs hook in the specRunner cannot.  These ones could be parallelized by using distinct file names so as not to clobber one another.

Adjusted some doc strings to fix some sphinx warnings.

Added tests for quota and assetstore policies for both server and web client.

Format sizes the same way in javascipt and python.

Offer to repick files from the upload widget if it fails due to space.

Moved testUpload to the general javascript testUtils so that more than one test can use it.  Finished tests for quota plugin.