Some changes to improve builds.
phantomjs occasionally fails to load javascript files.  This appears to be a known issue: https://github.com/ariya/phantomjs/issues/10652.  Retry several times if it looks like this has occurred.

I added code to better detect this condition.

Changed how the mock smtp loop is performed so as to not be using up 100% CPU.