Unset PYTHONPATH's before executing packaging tests
When PYTHONPATH contains the repository directory, girder-install
will detect the repository source as the install path.  Running
the tests reinstall plugins and web source and remove some files
that don't go in the distribution tarball.  This takes the
heavy handed approach of removing the PYTHONPATH variable entirely,
but for the purpose of these tests it probably doesn't matter much.