Improve modal-to-modal transition.
When we transition from one modal dialog to another, the old one still has some effect on the javascript up until the transition time ends (300 ms by default).  If we dismiss a second modal dialog within that time, sometimes bootstrap doesn't actually dismiss the second dialog properly (the backdrop is removed but not the dialog).  This occasionally resulted in a test failure.  I now remove the fade animation, so that bootstrap removes the old dialog synchronously, then put the animation back if it was there so that subsequent dialogs behave as expected.