Fixes #639. Cookie-based authentication options for route handlers
By default, the token value in the cookie is not sufficient to
authenticate to Girder (to protect against CSRF). There are a few
read-only endpoints (i.e. not vulnerable to CSRF attacks) that
make sense to open up to cookie-based authentication to support
user agent (browser) tools that make it infeasible to pass the normal
Girder-Token custom HTTP header since they are not requested with
normal XHR mechanisms. Namely:

* The notification stream endpoint, since in our web client it is
requested with an EventSource that does not allow specification of
headers in its API at this time.
* The download endpoints, since we want those links to download
as attachments to the user's filesystem and therefore have no
good way of specifying headers.

By allowing cookie auth for these endpoints, we also avoid placing
secure token values into query strings that might end up in
access logs on the server. Additionally, this allows us to simply
use the href property for download direct links, making it easy
for users to copy and disseminate direct download links, even to
private data.