Make all PyMongo "find" queries not timeout
A PyMongo cursor will always be deallocated by the server when it is
exhausted or when it is closed by the client (typically by going out of
scope). This change will ensure that cursors held open longer than
10 minutes will not be automatically closed by the server prematurely,
which could affect long-running batch jobs.

See: http://docs.mongodb.org/manual/core/cursors/#closure-of-inactive-cursors

Developers are still advised to manually close cursors that will not
quickly go out of scope, to ensure that both client and server-side
resources are freed as soon as possible.

Thanks to David Manthey for bringing up this topic.