Make renderMarkdown more generally useful
In the covalic plugin I have a place where I render some
markdown into a popover, and it's easiest to set the HTML when
instantiating the popover in javascript rather than placing
it in an element. For that case, and other cases like it that
may pop up, I've added a mode to renderMarkdown that just
parses and returns the HTML rather than putting it into an
existing element.