Add item&folder callbacks plus refactor to python client
Allows a caller of the python client code to set callbacks for folder
and item uploads.  The upload function and many of its helpers were
moved from the GirderCli to the GirderClient, as the Cli should be
concerned only with wrapping the GirderClient, and the GirderClient
should expose all methods that are desirable to be called programmatically.