Isolate tmp dir in python client lib_test
Testing of the folder and item upload callbacks used the same tmp
dir as the cli test, which could be run concurrently.  A race condition
could occur where the lib test snapshotted the tmp dir tree, then uploaded
the tree, and the cli test running concurrently modified the files in
the tmp dir tree in between the snapshot and upload, resulting in the lib
test failing as it saw a different set of files between snapshot and upload.