Several changes to make deployment easier.
The swagger api pages would only work if the static_root was absolute, because this was otherwise treated as relative to the api path.  I've added a new api_static_root parameter, which now allows all of the paths to be relative.  This makes it easier, for instance, for proxies to get to these locations.

The api_root is placed in the swagger page as well, as this also also the sub-describe calls to use a relative path instead of the entire query path.

Fixed the swagger throbber using css to override its path.  This still throws a 404 in the browser's console, when the original location is requested, but the throbber is properly displayed.  This fixes issue #446.

Add documentation that you don't need to specify a tools.proxy.base or tools.proxy.local if the proxy server adds X-Forwarded-Host.

Expand the user path element when parsing the log root, not just in the constants.  This allows a conf file to use ~ in the log root.