Correct emptyLayoutSpec
Previously emptyLayoutSpec was not using Jasmine's expect correctly,
this change corrects this.  This change further changes the empty layout
tests to test that the proper classes are added and removed from
the g-app-body-container, rather than testing for specific css property
values, which is more consistent with the current implementation and any
future expected changes.  Finally, there was a case where the default
layout was reapplied but the empty layout class had not been removed,
this test reveals that error.