Move install instructions for development into the same page
There's no need to clutter the server installation documentation page with
steps that are specific to developers. This moves some of these steps into
the development documentation page.

This also re-adds the installation of grunt and grunt-cli as global tools,
but makes it specific to developers, who may need to run grunt directly.