Update nginx deployment documentation to use correct header settings
The "X-Forwarded-Server" header is not used by Girder or CherryPy, nor is
it part of any standard. Apache automatically sets this header as a
redundancy, in case Apache's "ProxyPreserveHost" directive is used, but
there is no need for this to be set in nginx.

The "Host" header should be updated to use have the hostname of the proxy
server, so that Girder / CherryPy clearly knows that it sits behind a
proxy. All of the information from the original "Host" header is available
in the "X-Forwarded-Host" header. It's also important to explicitly set
the "Host" header, as HTTP/1.1 requires it to be present and the client
may have failed to set it.

The "X-Forwarded-Host" header should be set from nginx's "$host" variable,
rather than "$http_host". "$host" has the benefit of being able to fall
back to use the nginx's configured "server_name" if the client fails to
provide a "Host" header.