Remove unnecessary options from the proxy deployment documentation
If a proxy server is configured correctly, the "tools.proxy.base" and
"tools.proxy.local" options do not need to be set, and doing so makes a
Girder server less flexable.

Many people simply copy-paste portions of the suggested configuration
files from the documentation, so moving these options to a seperate
optional section will provide a cleaner configurarion for some users.