Make models always expose the '_id' field
All current models that use the filter capability add the '_id' field
to their list of exposed fields for all accesses. To ensure that the
important '_id' field is easily availble in new models, it is now part
of the default set of exposed keys.

It is still possible to call "Model.hideFields" if necessary.