Remove useless "models" property from Swagger resource descriptions
The "models" field in a Swagger resource description should contain a
detailed listing of the properties in a model. At the moment, Girder has
no mechanism to expose this information (though it might come via the
Model._filterKeys attribute). Without this information, there is no reason
to include a "models" field in resource descriptions. For more information,
see:
https://github.com/swagger-api/swagger-spec/blob/master/versions/1.2.md#52-api-declaration
https://github.com/swagger-api/swagger-spec/blob/master/versions/1.2.md#527-model-object

The empty "models" variable was first added by commit
a54c4337e6059519c91fb7c60d6e4441d55c2539 , but with no mechanism to ever
set or use it.

Later, commit cef7f4eac5e98987b99efa84068cc78e569ce937 finally started
returning the "models" property with the resource description endpoint,
but even assuming that the variable was set correctly, the code returns the
same thing for each seperate resource type. The "addModel" function isn't
tested or used anywhere, and only serves to set values in an
already-misused variable.

It's probably best at this point to consider this to be cruft, and
remove it all. Future work could attempt to define Model descriptions via
Swagger's specification, but this is nowhere close.