Make api.docs.addModel require a resource type
Per the Swagger specification, each model should be applied only to a
specific resource or set of resources. More rationale is provided in issue
#958

This commit changes the undocumented, but public, addModel function to
require each new model to specify a resource type which it will be
associated with. The same model may be added to multiple resources via
repeated calls to the function. This also updates the API resource
description endpoint to only return models which are associated with
each given resource type.

Closes #958.