Improve security of temporary access flow
* Refs #984.
* Fixes #955.

Before, when we created temporary access tokens, we also made them
have the scope of normal user auth tokens, and we also didn't
delete them after the user changed their password. Now, the temporary
token only grants the temporary auth scope, and it can be exchanged
for a normal auth token via the existing route for checking the
temp token. Also, it can now only be used once to change the
password, and is then deleted from the system.