Fix large zip files.
If a zip file contained multiple files and one of them started at an offset >= 2Gb from the start of the file, the zip footer would fail to be created.

Also, zip filenames with non-ASCII characters could fail to be created due to string conversions.

The test has been updated to detect both of these conditions.