Add skeleton for WADO-RS, QIDO-RS and STOW-RS API
The list of API endpoints are currently showing in swagger but that are
not implemented.

There are currently three resource objects named "dicomStudies", "dicomSeries"
and "dicomInstances" that add endpoints for WADO-RS, QIDO-RS and STOW-RS.

Note also that there is a function named "_reroute()" that was an
experiment to redirect calls to an existing instance providing similar
endpoints with the advantage of providing authentication.