Fix quota issues.
There were two issues.  If a user or collection never had a quota set on it, default quotas were not enforced.

In the web client, if logged in as a non-admin user and viewing the quota for a user or collection that used a default quota, it would show 'Unlimited' instead of the appropriate quota.

Add tests that would have failed before and now work.

Added a line in jsstyle.cfg to disable requireVarDeclFirst, a rule that is added is jscs 2.0 (for when we upgrade to that version).