Fixed a python 3 issue.
In python2, dict.keys() returns a list.  In python3, it returns an iterator, which means that the idiom of using 'for key in dict.keys()' to iterate through a non-mutating list where statements within the for loop no longer works in python3.  The new idiom seem to be 'for key in list(dict)'.