Move Webroot and Description (Swagger root) templates to separate files
This will make future maintenance of these files easier, as they are no
longer string literals constrained by Python style restrictions.

Since the new implementation stores the HTML string as an instance
variable, instead of a class variable, it is possible to instantiate
multiple Webroots with different templates in the same Girder instance.

This also slightly changes whitespace in the HTML template files, to
improve their style once they are actually sent to a client.