Make model creation an atomic database operation
In several places, a partially-created model is implicitly saved before
the end of the "create..." function. This creates several problems.

First, if the code terminates at a later part of the "create..." function,
the new database document will be left in an incomplete state.

Second, the premature save operation causes the "model.{}.save.created"
event to be fired before the "create..." function has finished. This causes
callbacks to receive an incomplete model.

Most significantly, since the remainder of the "create..." code is still
executed unconditionally after the "model.{}.save.created" callback
completes, it may undo some of the intentional functionality of the
callback. For example, a callback that sets a new user's "public"
attribute to False would have this change undone.