Add a api_url setting.
When behind a proxy, swagger doesn't work because of how we get the api url.  This now has a setting to override it.  If set to ".", the url will look ugly (as it has an extra . in the path, but will work both inside and outside of the proxy.  If set to "https?://(proxy host and port)/(proxy path)/api/v1", it will look nice and work through the proxy, but not locally.

A better technique would be to modify swagger-ui, but that seems heavy-handed for now.

I need to add tests for this and update the documentation, but wanted some feedback first.