Redesign `girder-install plugin` subcommand
The main improvement here is that, in addition to copying the plugin
into girder's package path and installing the pip reqs, this also
re-runs "npm install", which should build any web client targets
for the plugin. I also changed it so that rather than pointing to
a directory of plugins, you point it at a single plugin at a time.
Eventually I can see this being extended to a list of plugins,
which is probably a more familiar CLI format.

Also expanded the assertions of our packaging tests.