Add widgets to input and modify date/times
This commit adds two widgets: DateTimeWidget and DateTimeRangeWidget.

DateTimeWidget provides a text input field to specify a date/time. The user
chooses the date/time using a popup picker.

DateTimeRangeWidget provides text input fields to specify a range of
dates/times.  That is, the first field specifies "from" and the second field
specifies "to." The user chooses each date/time using a popup picker.

The widgets are implemented using the "eonasdan-bootstrap-datetimepicker"
module.

This commit also adds the Bootstrap 3.1.1 glyphicons font because
eonasdan-bootstrap-datetimepicker requires the font for its picker.