Fixes #1098. Move model loading behavior in loadmodel
This change fixes the breakage of the sphinx build on RTD.
The loadmodel decorator was actually loading the model during
__init__ (which was called at module import time). We move this
behavior into __call__ (which is called at request time). This
should fix the sphinx build on RTD, which was breaking by trying
to connect to a nonexistent mongodb on localhost.