Add corsAuth option for saving auth token in CORS env
Users can set `girder.corsAuth = true` in their app prior to
instantiating their girder app, which will allow the token
to be saved on the origin domain even if girder is on a
different origin.

This also introduces a new handy variable `girder.currentToken`
that stores the currently logged-in user's token.