Create a filtermodel decorator
This wrapper will filter a returned model or list of models
and is meant to be used to wrap REST route handlers that return
models. This commit also contains cleanup of the ``filter()``
overrides that are in place in the folder and item models to
preserve backward compatibility with the old kwargs prior to
the refactoring of filtering functionality.