Timeline widget (first pass)
This is a general purpose widget for rendering two types of primitives
on a time line: segments and points. Segments represent time spans,
points represent individual points in time.

I've added an instance of this widget to the job details view to
visualize the job lifecycle and provide tooltips for the duration
of each step. Segments are used to represent the time in which a
job was in a given status.