Add 'files' processing logic
Allows for ensuring a list of files are associated with an item.  This
this is idempotent (i.e.,  files will be DELETED that are not in the
list,  files not currently associated with the item will be uploaded).
File equality is based on os.path.basename() and file size as reported
by the girder API and os.path.getsize().