Properly treats ENABLEDPLUGINS as a multival arg
Before this was invoked as
add_web_client_test(...
ENABLEDPLUGINS "foo" "bar" "baz"
..)

This caused the actual environment variable to be set as whatever the
first value was, foo in this case.

Quoting the value doesn't work, so we replace the semicolons with spaces
in the list so the environment variable is space separated. This works
with Girder since it uses str.split to separate its arguments.