Remove describeResource
In Swagger Spec 2.0, the resource listing and API declaration are combined. The
entire API description is contained in a single document.

Previously, the /describe endpoint returned the Resource Listing and the
/describe/{resource} endpoint returned the API Declaration for a particular
resource. Now, the /describe endpoint returns the complete API.

This commit removes the /describe/{resource} endpoint, which is now unnecessary.