Sort endpoints on swagger-ui front-end
In Swagger Spec 2.0, the Paths Object that contains the routes is a dictionary
and therefore has no concept of order. Previously, the server sorted the
endpoints in a list when generating the API description. To maintain the earler
sorting order on the swagger-ui page, this commit reimplements the sorting using
swagger-ui's 'operationsSorter' parameter.