Update describe tests for Swagger Spec 2.0
Tests for the /describe/{resource} endpoint are removed because that endpoint no
longer exists.

The route order test is no longer valid since the spec contains the routes in a
dictionary; the swagger-ui front-end now handles sorting. In its place is a test
that checks whether the expected routes and methods exist. To make any failures
trivially explicit, the test replaces looping over the 'Routes' list with
individual assertions.

The model test is updated to use the 'definitions' field.