Make access.cookie only support force flag as kwarg
I prefer this simplicity since we have to dynamically try to
disambiguate between a raw decorator vs. a function that is
supposed to return a decorator based on the arguments passed in;
that distinction becomes much less tricky if the latter case
only supports passing flags via kwargs. Otherwise there is a
theoretical case that is impossible to disambiguate.