Workaround for the grunt-fontello task.
The grunt-fontello task mutates the fontello config file.  If this file has a name entry that is an invalid or expired fontello session ID, the task fails.  To work around this, copy the fontello config file to an other directory and have the grunt-fontello task use that.  This means that we always will get a new fontello session, and the task will avoid the condition of having a bad session.  This also avoids mutating the source file.