When importing a module by entry point, make sure it is available via girder.plugins[modulename] so that its submodules are always reachable in a consistent fashion.
Read a plugin.json or plugin.yml file from the packaged plugin rather than requiring this information to be duplicated at the entry point with the @plugin decorator.  This expects the plugins file within the module to be a girder/plugin.json or girder/plugin.yml.

Note that plugins must not import other plugins at the time that the module with the entry point is imported, as checking the @plugin decorator OR reading module files (such as plugin.json) import the module BEFORE module dependencies have been sorted and loaded.

The most useful feature of this branch is making sure the plugin is in the appropriate spot of the sys.modules namespace so that it can import its own modules.  This isn't currently tested, as it will require installing and enabling a plugin that isn't in the plugins directory.