Add item licenses plugin
When the item licenses plugin is enabled, items gain a 'license' field.

In the web client, users can select a license from a dropdown when uploading an
item, creating an item, or editing an item.

The list of licenses is configurable in the plugin settings. The default
licenses include the some of the "popular" open source licenses listed on
https://opensource.org/licenses/category, the Creative Commons licenses listed
on http://creativecommons.org/licenses/, as well as CC0, Public Domain Mark, and
All Rights Reserved.

A future extension to this plugin could be to integrate the license selector
widget from https://github.com/ufal/public-license-selector. To make such
integration easier, the license names in this plugin exactly match the names in
that widget.