Instead of just printing to stdout, also send information to logs.
Except for the install utilities, all print statements now go through the logprint function.  This still prints the value to stdout with the appropriate terminal color.  It also send the information to the configure logs (if any).  This simplifies printing in color, as instead of print(TerminalColor.info(msg)), one can use logprint.info(msg).  It also adds exception printing support and a few additional log levels.

This resolves issue #602.