Remove user access checks from _pruneOrphans in consistancy check
A user may have access to a sub-folder that is within a parent folder or
collection that the user does not have access to (this is how "unlisted"
folders can be implemented). Thus, these sort of sub-folders are not
technically 'orphans' and definitely should not be removed.

As a practical matter, the system consistancy check may only be run by a
site admin user, who will have access to everything.