add more widget options and client-side filtering
Adds options to the HierarchyWidget and other
related widgets that allow client-side filtering
of the items/folders that are to be displayed.
Also added a "folder selection mode".

In order to get the filtering working, I had to
extend the base Collections class with additional
filtering logic.  As much as possible, I tried to
keep the filtering completely transparent to any
subclasses.

The addition of transparent filtering gets tricky
when considering the case of pagination.  The
fetch logic now makes multiple round-trips as
needed to fill up an entire "page", filtering as
it goes.  The trickier issue, however, was in
dealing with going backwards
(fetchPreviousPage()).  The best solution I came
up with was to maintain a stack of offsets from
which pages where successfully fetched.  In any
case, all the functional changes should be
entirely opt-in, and should still work same as
before by default.

 - Added new per-instance attribute to
   Collections: "filter".

   - Set to a function to perform client-side
     filtering of fetched collection items.

     ```
     /* set this attribute before fetching data */
     myFolderListView.collection.filter = (
         (model) => model.name !== 'test');
     ```

 - Added new options to FolderListWidget

   - folderFilter: collection filter function for
     folders

 - Added new options to ItemListWidget

   - itemFilter: collection filter function for
     items

 - Added new options to HierarchyWidget

   - itemFilter: collection filter function for
     items and folders

   - selectFolder: set to true to enable folder
     selection mode

   - onFolderSelect: callback triggered when
     selecting a folder in folder selection mode

     - selectFolder and onFolderSelect must both
       be set to enable folder selection mode.  In
       this mode, a button is created at the top
       of every folder view with the text "Select
       this folder/user/collection."  Useful for
       prompting a user to select a girder folder.