Stop using an exposed webroot for static files
Since the BaseWebroot is mounted using the staticdir tool, it uses its
own dispatcher and doesn't need its own webroot, see the docs for
cherrypy.tree.mount.

Since it was exposed but didn't have any retrieval methods (i.e. GET),
it was expecting to be callable when the static handler wasn't fulfilled
(file not found) and throwing an exception (and a 500).