Force vagrant-ansible examples to always build assets
This is important because vagrant mounts the Girder directory and the
defaults for the vagrant-ansible examples are to not git update/force
changes. This means when provisioning, git will never retrieve new files
and a new girder-install web would never be
triggered. girder_always_build_assets forces the assets to always be
built on provision, regardless of the results of the git step.