Make logs more configurable.
Allow setting the size of log files, the number of logs kept, and the log level in the config file.  Also provide an option to suppress all output to stdout.

Make it easier to log cherrypy accesses to the info log.  You can, in the [global] section, add
  log.screen: None
to disable stdout access logging, and
  log.access_file: (path to info log)
to log accesses to the info log (however this doesn't honor rotation properly).

By adding a log_access parameter in the [logging] configure section, this can be controlled more easily.  This can take a list, tuple, set, or single string from the list ('screen', 'info').  An empty list will not log accesses at all (pedantically cherrypy logs to the null handler in any case).  When logging to the info log, our usual format is not applied, as this adds the time stamp a second time (though with milliseconds) and the INFO level tag, which seems unneeded for accesses.

This has the virtue that you don't need to repeat the log file name or path (whereas to use the [global] log.access_file, you do), and that logs are rotated as per our rules.

When reading logs via the endpoint, handle recent rotations.

Add an endpoint for setting log levels.

Convert some bare print statements to logprint statements.  Except for installation we should make it a habit of NEVER leaving print statements in the code -- logging should always preferred.  That way, information can be accessed from the system log endpoint.