girder_client: Revert to checking server version for selecting upload strategy
This commit partially reverts cb05eb4 (girder_client: Improve detection
of non-multipart upload support). Indeed, the "chunk" parameter was
removed from the params reported by Swagger because Girder autoDescribeRoute
is insufficiently powerful to handle a param type that could be either in
the body or in the params, but in the future this could change.