Protect unspecified, null-valued keys from deletion
If a null value is installed in the metadata using `allowNull=true`,
then a subsequent metadata record is modified without using
`allowNull=true`, the logic will sweep the item metadata for null values
and strike them from the metadata. This "non-local" effect is
undesirable, but the fix is simple - instead of sweeping the *metadata*
for null-valued keys to strike, simply sweep the *input* instead.