Improve error checking when user attempts to upload a directory
Add functionality to the upload widget to check whether every dropped
item is a file. Display an error message when the check fails.

Note that the implementaion uses the DataTransferItem.webkitGetAsEntry()
method, which has not yet been standardized, but first tries to use the
suggested future standard method name getAsEntry(). See
https://developer.mozilla.org/en-US/docs/Web/API/DataTransferItem.