Remove the Webpack url-loader
The 'url-loader' provides a wrapper around 'file-loader', so the two
loaders should not be chained together. This bug was not discovered, as
Girder was not loading any '.woff', '.ttf', or '.svg' files with Webpack.

Rather than fix the chaining bug, this just removes the url-loader, as
its use is probably a premature optimization. Using url-loader will
theoretically improve first-load performance, but makes it more difficult
for user agents to cache assets and will interfere with the effectiveness
of HTTP2 in the future.