Always include API version in system/version response
This fixes a bug when the girder web client does not exist. In that
case, the API_VERSION value is set to null, even though there's no
reason for us not to report it since it's known in the backend.

This commit changes the behavior by always ensuring API_VERSION
is set to the package version, regardless of the existence of
girder-version.json.