BUG: Don't display items containing multiple files in item_preview plugin
Since an item containing multiple files is going to be download as a zip file,
and the item_preview do not support a display inline of a zip file, it produces
a display error.

Then, one solution is to check all items and remove the ones containing
multiple files from `supportedItems`.

Cons:
* Do not preview:
  - supported item containing multiple supported files
  - item named `example` containing a supported file (indeed, for now,
    an item is defined as supported if its name contain .png at the end
    for instance)
* Error still shown:
  - if an item considered as supported (with .png at the end of its name for
    instance) and the file inside is not supported (a .txt file for example)