Prevent deleting a file that is being uploaded.
If an upload occurs where the file has the same hash as an existing file, we don't duplicate the file.  There was a window of time where if a delete occurred during the upload, we would detect that the file already existed, but then allow the delete to occur.  This fixes this by checking to see if there are any pending uploads with the file and using a lock to ensure that the check and deletes cannot conflict with one another.

This fixes issue #1999.