Wait for bootstrap transitions.
We use bootstrap transitions on dialogs.  These are meant to trigger the appropriate browser event, which, if it doesn't fire in the expected amount of time, also uses a window timeout as a fallback to trigger the event.

Unfortunately, there isn't currently any official way in Bootstrap 3 to know if you are in a transition state, and some of our tests rely on the previous activities being completed before starting the next activities.   Bootstrap 4 is adding an internal flag to determine if it is in a transition, but it has not been released yet.

This adds our own testing event for transition end, ignoring the browser's css transitionend event.  It allows a flag to specify that we are in a transition, and therefore allows waitForDialog and waitForLoad to detect if we are still in a transition.