Don't override browser css transition end.
I originally did so because, since the browser css transitionend message being unreliable meant that nothing should be able to usefully bind to them.  By using our own event, it was guaranteed to always behave in an expected way.  As @jbeezley mentioned, there may be side effects to doing this, since something might bind to those css events.  This change addresses that.