GridFS assetstore speed improvements.
Speed up GridFS deletes by adding the right index.

We check if a gridfs chunk is used by querying for the chunkUuid.  There was no index on this, so deletes were very slow with even a modest number of GridFS files.

Record if we have used a Mongo connection recently.  If so, don't ensure indices on it.  On my local system using a remote mongo GridFS server, ensuring indices that already exist consumes an average of around 6 ms.