Adjust boto S3 connection to produce better error messages
When validate=True was passed, the error message for an invalid
service url was a 301 with no further information. By setting
validate=False and manually calling get_all_keys(maxkeys=0) we
effectively do the same thing, but the log produces an error message
hinting to a problem with the region name. This is the approach
suggested in the docstring of boto.s3.connection.get_bucket.