Drop test databases at the end of the test.
Currently, we reference a mongo database, usually one per test.  Some tests that use gridfs reference additional databases.  With mongo, storing values in a database that doesn't yet exist creates it.  We try to drop the databases as we start tests, but there are conditions where we read the database before we drop it (such as determining the static root).  By removing the databases at the end of the tests, this eliminates behavioral changes on different test runs depending on where the database persisted or not.  Also, it avoids cluttering mongo with databases when one isn't running tests.

This can be overridden with EXTRADEBUG=keepdb.

If EXTRADEBUG includes 'newdb', a failure will be triggered if the database already exists (so running tests using EXTRADEBUG='keepdb newdb' twice should always result in a failure).