Remove custom caching from webroot HTML/vars
This commit makes Girder do things in a more expensive manner by
default to allow for a more generalized caching
solution. Particularly, the webroot was caching which plugins were
enabled as well as the entire indexHtml in memory. This led to issues
where it would be necessary to listen for settings to change to know
when to invalidate the cache. Furthermore, this prevents the ability
to effectively leverage caching in a distributed installation of
Girder.

As of this commit indexHTML is no longer being cached, however the
enabled plugins setting is (from a prior commit), and the remaining
variables are dictionary lookups already.