Add a cookiecutter template for a Girder plugin
This PR introduces a template for a popular python scaffolding tool
called cookiecutter: https://github.com/audreyr/cookiecutter

The main goal of such a template is to demonstrate best practices for
building a Girder plugin, and pointing the user to more in depth
resources for the various parts of the code. This means as new
features and practices of coding are developed for Girder, they must
be reflected in the template.

The built plugin demonstrates the following:
- Writing a client side test
- Writing a server side test
- Binding to server-side events
- Working with server-side models
- Adding a REST endpoint
- Adding a custom setting with a default and validator
- Adding a custom configuration view for changing said setting
-- Exercising the REST API from JavaScript
-- Listening to events in JavaScript
-- Rendering a Pug template
- Wrapping an existing client-side view

To make use of this, perform the following commands from Girder root:
pip install cookiecutter
cookiecutter devops/cookiecutter-girder-plugin
girder-install plugin -s your-cookiecutter-plugin